% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.connect_operations.R
\name{create_user}
\alias{create_user}
\title{Creates a new user account in your Amazon Connect instance}
\usage{
create_user(Username, PhoneConfig, SecurityProfileIds, RoutingProfileId,
  InstanceId, Password = NULL, IdentityInfo = NULL,
  DirectoryUserId = NULL, HierarchyGroupId = NULL)
}
\arguments{
\item{Username}{The user name in Amazon Connect for the account to create.}

\item{PhoneConfig}{Specifies the phone settings for the user, including AfterContactWorkTimeLimit, AutoAccept, DeskPhoneNumber, and PhoneType.}

\item{SecurityProfileIds}{The unique identifier of the security profile to assign to the user created.}

\item{RoutingProfileId}{The unique identifier for the routing profile to assign to the user created.}

\item{InstanceId}{The identifier for your Amazon Connect instance. To find the ID of your instance, open the AWS console and select Amazon Connect. Select the alias of the instance in the Instance alias column. The instance ID is displayed in the Overview section of your instance settings. For example, the instance ID is the set of characters at the end of the instance ARN, after instance/, such as 10a4c4eb-f57e-4d4c-b602-bf39176ced07.}

\item{Password}{The password for the user account to create. This is required if you are using Amazon Connect for identity management. If you are using SAML for identity management and include this parameter, an \code{InvalidRequestException} is returned.}

\item{IdentityInfo}{Information about the user, including email address, first name, and last name.}

\item{DirectoryUserId}{The unique identifier for the user account in the directory service directory used for identity management. If Amazon Connect is unable to access the existing directory, you can use the \code{DirectoryUserId} to authenticate users. If you include the parameter, it is assumed that Amazon Connect cannot access the directory. If the parameter is not included, the UserIdentityInfo is used to authenticate users from your existing directory.

This parameter is required if you are using an existing directory for identity management in Amazon Connect when Amazon Connect cannot access your directory to authenticate users. If you are using SAML for identity management and include this parameter, an \code{InvalidRequestException} is returned.}

\item{HierarchyGroupId}{The unique identifier for the hierarchy group to assign to the user created.}
}
\description{
Creates a new user account in your Amazon Connect instance.
}
\examples{

}
