% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.connect_operations.R
\name{list_routing_profiles}
\alias{list_routing_profiles}
\title{Returns an array of \code{RoutingProfileSummary} objects that includes information about the routing profiles in your instance}
\usage{
list_routing_profiles(
  InstanceId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
\arguments{
\item{InstanceId}{[required] The identifier for your Amazon Connect instance. To find the ID of your instance, open the AWS console and select Amazon Connect. Select the alias of the instance in the Instance alias column. The instance ID is displayed in the Overview section of your instance settings. For example, the instance ID is the set of characters at the end of the instance ARN, after instance/, such as 10a4c4eb-f57e-4d4c-b602-bf39176ced07.}

\item{NextToken}{The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.}

\item{MaxResults}{The maximum number of routing profiles to return in the response.}
}
\description{
Returns an array of \code{RoutingProfileSummary} objects that includes information about the routing profiles in your instance.
}
\examples{

}
