% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.connect_operations.R
\name{update_user_routing_profile}
\alias{update_user_routing_profile}
\title{Assigns the specified routing profile to a user}
\usage{
update_user_routing_profile(RoutingProfileId, UserId, InstanceId)
}
\arguments{
\item{RoutingProfileId}{The identifier of the routing profile to assign to the user.}

\item{UserId}{The identifier for the user account to assign the routing profile to.}

\item{InstanceId}{The identifier for your Amazon Connect instance. To find the ID of your instance, open the AWS console and select Amazon Connect. Select the alias of the instance in the Instance alias column. The instance ID is displayed in the Overview section of your instance settings. For example, the instance ID is the set of characters at the end of the instance ARN, after instance/, such as 10a4c4eb-f57e-4d4c-b602-bf39176ced07.}
}
\description{
Assigns the specified routing profile to a user.
}
\examples{

}
