% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.connect_operations.R
\name{update_user_security_profiles}
\alias{update_user_security_profiles}
\title{Updates the security profiles assigned to the user}
\usage{
update_user_security_profiles(SecurityProfileIds, UserId, InstanceId)
}
\arguments{
\item{SecurityProfileIds}{The identifiers for the security profiles to assign to the user.}

\item{UserId}{The identifier of the user account to assign the security profiles.}

\item{InstanceId}{The identifier for your Amazon Connect instance. To find the ID of your instance, open the AWS console and select Amazon Connect. Select the alias of the instance in the Instance alias column. The instance ID is displayed in the Overview section of your instance settings. For example, the instance ID is the set of characters at the end of the instance ARN, after instance/, such as 10a4c4eb-f57e-4d4c-b602-bf39176ced07.}
}
\description{
Updates the security profiles assigned to the user.
}
\examples{

}
